pyproject.toml
[project]
name = "family-planning-data-extraction-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncer>=0.0.11",
    "connectorx>=0.4.4",
    "fastapi[standard]>=0.123.10",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "polars>=1.35.2",
    "psycopg2>=2.9.11",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "shortuuid>=1.0.13",
    "sqlalchemy>=2.0.44",
]
